faet: add attachment to announcement

+++ services/announcementService.js
@@ -1,5 +1,7 @@
 const mongoose = require('mongoose');
 const AnnouncementModel = require('../models/announcementModel');
+const fs = require('fs');
+const path = require('path');
 
 // Get all announcements
 const getAnnouncements = async () => {
@@ -25,7 +27,7 @@ const getAnnouncement = async (id) => {
     return announcement;
 };
 
-const createAnnouncement = async (body) => {
+const createAnnouncement = async (body, file) => {
     try {
         const {
             title,
@@ -35,7 +37,11 @@ const createAnnouncement = async (body) => {
         const newAnnouncement = new AnnouncementModel({
             title,
             description,
+            fileURI: file ? file.filename : null,
+            fileName: file ? file.originalname : null,
         });
+
+        console.log(newAnnouncement);
 
         const savedAnnouncement = await newAnnouncement.save();
         return savedAnnouncement;
@@ -64,10 +70,39 @@ const deleteAnnouncement = async (id) => {
     if (!mongoose.Types.ObjectId.isValid(id)) {
         throw new Error('Announcement not found');
     }
+    // Find announcement by id and delete attachment
+    const result = await AnnouncementModel.findById(id);
+    if (!result) {
+        throw new Error('Announcement not found');
+    }
+    if (result.fileURI) {
+        await deleteAttachment(result.fileURI);
+    }
     // Find announcement by id and delete
     const announcement = await AnnouncementModel.findByIdAndDelete(id);
     if (!announcement) {
         throw new Error('Announcement not found');
+    }
+};
+
+const deleteAttachment = async (fileURI) => {
+    try {
+        // Resolve the full path to the file
+        const filePath = path.join(__dirname, '..', 'uploads', fileURI); // Ensure './uploads' matches your file storage directory
+        console.log(`Attempting to delete: ${filePath}`);
+
+        // Use the resolved file path in fs.unlink
+        fs.unlink(filePath, (err) => {
+            if (err) {
+                console.error(`Error deleting attachment ${filePath}:`, err);
+            } else {
+                console.log(`Deleted attachment ${filePath}`);
+            }
+        });
+
+        return { message: 'Attachment deleted successfully!' };
+    } catch (error) {
+        throw new Error(error.message);
     }
 };
 
@@ -77,4 +112,5 @@ module.exports = {
     createAnnouncement,
     updateAnnouncement,
     deleteAnnouncement,
+    deleteAttachment,
 };
